2017-05-29 Har fixat så att collision fungerar nu, det är inte buggigt. Det tar alltså inte bort fel bil från listan, utan det är alltid rätt bil. Däremot fick jag spelet att krasha när hastigheten på Ally och enemy är på Position += 1 Det ger ett error på CheckAllyEnemyCollisions metoden. Att i inte får ligga utanför index eller att index är för lågt. Score när player passerar en bil räknas ibland inte alltid ett "Score" med utan hoppar över en bil ibland. Detta är för att hastigheten på bilen är tex 8, vilket adderar 8 på position. Eftersom poäng räknas när player Y positionen är samma som tex enemy så är inte alltid Y positionen samma även fast de passerar varandra.

+++ slutprojekt_programmering2/slutprojekt_programmering2/Car.cs
@@ -97,7 +97,7 @@ namespace slutprojekt_programmering2
                null,
                Microsoft.Xna.Framework.Color.White,
                Rotation,
-               new Vector2(Texture.Width / 2, Texture.Height / 2),
+               new Vector2((float)Texture.Width / 2, (float)Texture.Height / 2),
                new Vector2(1,1),
                SpriteEffects.None,
                0);
